+++ tests/capture/test_capture_while_loop.py
@@ -216,6 +216,43 @@ class TestCaptureCircuitsWhileLoop:
         res_ev_jxpr = jax.core.eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, *args)
         assert np.allclose(res_ev_jxpr, expected), f"Expected {expected}, but got {res_ev_jxpr}"
 
+    @pytest.mark.parametrize("upper_bound, arg", [(3, 0.5), (2, 12)])
+    def test_while_and_for_loop_nested(self, upper_bound, arg):
+        """Test that a nested while and for loop is correctly captured into a jaxpr."""
+
+        dev = qml.device("default.qubit", wires=3)
+
+        def ry_fn(arg):
+            qml.RY(arg, wires=1)
+
+        @qml.qnode(dev)
+        def circuit(upper_bound, arg):
+
+            # while loop with dynamic bounds
+            @qml.while_loop(lambda i: i < upper_bound)
+            def loop_fn(i):
+                qml.Hadamard(wires=i)
+
+                @qml.for_loop(0, i, 1)
+                def loop_fn_returns(i, x):
+                    qml.RX(x, wires=i)
+                    m_0 = qml.measure(0)
+                    qml.cond(m_0, ry_fn)(x)
+                    return i + 1
+
+                loop_fn_returns(arg)
+                return i + 1
+
+            loop_fn(0)
+
+            return qml.expval(qml.Z(0))
+
+        args = [upper_bound, arg]
+        result = circuit(*args)
+        jaxpr = jax.make_jaxpr(circuit)(*args)
+        res_ev_jxpr = jax.core.eval_jaxpr(jaxpr.jaxpr, jaxpr.consts, *args)
+        assert np.allclose(result, res_ev_jxpr), f"Expected {result}, but got {res_ev_jxpr}"
+
 
 def test_pytree_input_output():
     """Test that the while loop supports pytree input and output."""
